bot/requirements.txt
pytest==6.2.5
pytest-asyncio==0.16.0
aioresponses==0.7.2
pytest-aiohttp==0.3.0
marshmallow-dataclass==8.5.3
aiobotocore==1.4.2
pyyaml==6.0
aio-pika==6.8.0
alembic==1.9.4
Mako==1.2.4
SQLAlchemy==2.0.4
greenlet==2.0.2
MarkupSafe==2.1.2
asyncpg==0.26.0
apispec==3.3.2
webargs==5.5.3
Jinja2==3.1.2
aiohttp==3.8.4
marshmallow==3.19.0
